フォルダ配置換え。テキストも保存するように変更
